feat: 메서드 일부 PresidentService로 분리
